Use an OBJECT library to support library tests
On Windows, the explicit module interface prevents us from calling the
non-exported functions in the library from tests. To better facilitate
testing of these functions without the need to explicitly list
dependencies of those functions (both internal and external), create an
OBJECT_LIBRARY with all of the library functions for use by the tests.

In comparison to the current implementation, this has the added benefit
of not compiling the source files multiple times, also meaning that the
compiled objects that we're testing are the same objects shipped in the
product.

Additionally, the `-fPIC` flag was dropped from the bundled PCRE flags
when we're not building a shared library. The `digest` test target was
also added to the `check` target dependencies.