Use SIGUSR1 to trigger a soft shutdown state
This new 'soft' shutdown state is used to trigger a shutdown without
disconnecting active proxy connections. In this state:
* Registration is reported as 'Busy' because there are 0 total slots
* The listening socket is shutdown, so no new connections can be made
* Existing users can continue to use the proxy as usual

Eventually, I'd like to be able to use SIGUSR2 to bring the proxy out of
this state, but I'll need to do a little refactoring to make that
possible. To best align with that goal, the current implementation will
put the proxy into a 'busy loop' to keep from shutting down completely.
Since the SIGUSR2 handler isn't done, the only thing that can happen
after a soft shutdown is a full shutdown (for now).