Make proxy_conn threads persist
The idea behind this change is to reduce the amount of work when a new
client is connected. Before this change, two threads were started
whenever a client connected, and another was started whenever a TCP
connection was opened on a proxy client's behalf.

This change makes all of those threads persist, but sleep whenever
they're not needed. This should make the connection process faster and
more reliable in exchange for a slightly larger overhead when no clients
are connected in the form of idle threads.