Attempt to give clients the same slot on reconnect
If a client disconnects and then reconnects before their previous slot
has been used by a different callsign, attempt prefer using the same
slot again.

A typical client will register with the EchoLink registration server
under the slot's external IP address. Other EchoLink users will download
the directory that contains that registration. If a client disconnects
and reconnects quickly, it would be nice to give them the same external
address so that other EchoLink users who attempt to connect to that user
using the information in the directory will find the client at the same
address. If we give them a new one, there other EchoLink user would need
to wait for the directory to be refreshed with the new address.