Drop HTML Help generation
I should have known bettern than to rely on an MSDN link staying around.
I can't find any replacement for the HTML Help Workshop, which is a
pretty old tool at this point. I'd guess I'll have to drop the CHM file
generation for now.

Also bump the GitHub upload action to the released v2 version now that
it is out of preview.