Drop clients from proxy_worker, not proxy_conn
A previous refactor moved the authorization process from the proxy_conn
to the proxy_worker, but did not introduce any mechanism to drop an
in-progress authorization before a client connection was assigned to an
available proxy_conn for normal use. Because ownership of the client
connection was also moved to the proxy_worker, it's safe to initiate the
connection dropping process from the proxy_worker instead of the
proxy_conn, which will cover the in-progress authorization case as well.

Fixes 7709624cbc9dbb15b1f479126f211164ada53314