Fix socket handle leak for outgoing TCP connections
If a proxy client sends a TCP_OPEN message while the proxy already has
an open connection, the proxy will leak the handle to the original
connection to process the new one.

After testing this scenario on the Java implementation, it appears to
send a TCP_CLOSE message to the client for the original connection and
then immediately open a new connection. It's hard to tell if this is
what the EchoLink client wants us to do, but taking the same behavior as
the Java proxy is certainly better than leaking the handle like we were
doing, so that's what this change does.